Add overwrite option and bugfix for pytest
- add overwrite parameter to avoid re-downloading if the file exists
- directly import urllib.request to avoid error when running pytest

+++ src/spicetools/queryutil.py
@@ -1,10 +1,12 @@
-import os
 from pathlib import Path
-import urllib
+from urllib import request
 
 
-def download_jpl_de(dename="de440s", output=None):
-    """Download JPL development ephemeris file
+__all__ = ["download_jpl_de"]
+
+
+def download_jpl_de(dename="de440s", output=None, overwrite=False):
+    """Download JPL development ephemeris file (intended to be used one time).
 
     Parameters
     ----------
@@ -15,10 +17,14 @@ def download_jpl_de(dename="de440s", output=None):
     output : str, pathlib.Path, optional
         Name of the output file path. If not provided, it will be saved into
         this package's `kernels/` directory.
+
+    overwrite : bool, optional
+        If `True`, overwrite the existing file if it exists. Default is
+        `False`, i.e., not download any file but returns the path to the
+        existing one.
     """
     if not dename.endswith(".bsp"):
         dename += ".bsp"
-    url = f"https://ssd.jpl.nasa.gov/ftp/eph/planets/bsp/{dename}"
 
     if output is None:
         # output = os.path.join(os.path.dirname(os.path.abspath(__file__)), "kernels", dename)
@@ -26,11 +32,15 @@ def download_jpl_de(dename="de440s", output=None):
     else:
         output = Path(output)
 
+    if output.exists() and not overwrite:
+        return output
+
     # Download the file
-    response = urllib.request.urlopen(url)
+    url = f"https://ssd.jpl.nasa.gov/ftp/eph/planets/bsp/{dename}"
+    response = request.urlopen(url)
+
     with open(output, 'wb') as f:
         f.write(response.read())
 
     print(f"Downloaded {dename} to {output}")
     return output
-
